+++ oap-server/oal-rt/src/main/java/org/apache/skywalking/oal/rt/parser/OALListener.java
@@ -21,10 +21,16 @@ package org.apache.skywalking.oal.rt.parser;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import org.antlr.v4.runtime.misc.NotNull;
 import org.apache.skywalking.oal.rt.grammar.OALParser;
 import org.apache.skywalking.oal.rt.grammar.OALParserBaseListener;
+import org.apache.skywalking.oap.server.core.analysis.ISourceDecorator;
+import org.apache.skywalking.oap.server.core.analysis.SourceDecoratorManager;
+import org.apache.skywalking.oap.server.core.analysis.metrics.LabeledValueHolder;
+import org.apache.skywalking.oap.server.core.analysis.metrics.Metrics;
 import org.apache.skywalking.oap.server.core.source.DefaultScopeDefine;
+import org.apache.skywalking.oap.server.core.source.ISource;
 
 public class OALListener extends OALParserBaseListener {
     private List<AnalysisResult> results;
@@ -105,6 +111,32 @@ public class OALListener extends OALParserBaseListener {
     public void exitFuncParamExpression(OALParser.FuncParamExpressionContext ctx) {
         current.getAggregationFuncStmt().addFuncConditionExpression(conditionExpression);
         conditionExpression = null;
+    }
+
+    @Override
+    public void enterDecorateSource(OALParser.DecorateSourceContext ctx) {
+        Class<? extends Metrics> metricsClass = MetricsHolder.find(current.getAggregationFuncStmt().getAggregationFunctionName());
+        if (LabeledValueHolder.class.isAssignableFrom(metricsClass)) {
+            throw new IllegalArgumentException(
+                "OAL metric: " + current.getMetricsName() + ", decorate source not support labeled value metrics.");
+        }
+        String decoratorName = ctx.STRING_LITERAL().getText();
+        String decoratorNameTrim = decoratorName.substring(1, decoratorName.length() - 1);
+        current.setSourceDecorator(decoratorNameTrim);
+        Map<String, ISourceDecorator<ISource>> map = SourceDecoratorManager.DECORATOR_MAP;
+        int currentScopeId = current.getFrom().getSourceScopeId();
+        if (currentScopeId != DefaultScopeDefine.SERVICE) {
+            throw new IllegalArgumentException("OAL metric: " + current.getMetricsName() + ", decorate source only support service scope.");
+        }
+        ISourceDecorator<ISource> decorator = map.get(decoratorNameTrim);
+        if (decorator == null) {
+            throw new IllegalArgumentException("OAL metric: " + current.getMetricsName() + " define a decorator: " + decoratorNameTrim
+            + ", but can't find it.");
+        }
+        int scopeId = decorator.getSourceScope();
+        if (scopeId != currentScopeId) {
+            throw new IllegalArgumentException("OAL Decorate Source, expect decorator scope id is: " + currentScopeId + ", but got: " + scopeId);
+        }
     }
 
     /////////////