+++ oap-server/server-storage-plugin/storage-banyandb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/banyandb/stream/AbstractBanyanDBDAO.java
@@ -19,6 +19,7 @@
 package org.apache.skywalking.oap.server.storage.plugin.banyandb.stream;
 
 import com.google.gson.Gson;
+import java.util.Arrays;
 import org.apache.skywalking.banyandb.model.v1.BanyandbModel;
 import org.apache.skywalking.banyandb.v1.client.AbstractCriteria;
 import org.apache.skywalking.banyandb.v1.client.AbstractQuery;
@@ -34,11 +35,13 @@ import org.apache.skywalking.banyandb.v1.client.TimestampRange;
 import org.apache.skywalking.banyandb.v1.client.TopNQuery;
 import org.apache.skywalking.banyandb.v1.client.TopNQueryResponse;
 import org.apache.skywalking.banyandb.v1.client.Trace;
+import org.apache.skywalking.oap.server.core.analysis.metrics.Metrics;
 import org.apache.skywalking.oap.server.core.query.type.KeyValue;
 import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingSpan;
 import org.apache.skywalking.oap.server.core.query.type.debugging.DebuggingTraceContext;
 import org.apache.skywalking.oap.server.core.storage.AbstractDAO;
 import org.apache.skywalking.oap.server.library.util.CollectionUtils;
+import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.storage.plugin.banyandb.BanyanDBStorageClient;
 import org.apache.skywalking.oap.server.storage.plugin.banyandb.MetadataRegistry;
 import java.io.IOException;
@@ -120,18 +123,28 @@ public abstract class AbstractBanyanDBDAO extends AbstractDAO<BanyanDBStorageCli
         }
     }
 
-    protected TopNQueryResponse topN(MetadataRegistry.Schema schema, TimestampRange timestampRange, int number,
-                                     List<KeyValue> additionalConditions) throws IOException {
-        return topNQuery(schema, timestampRange, number, AbstractQuery.Sort.DESC, additionalConditions);
+    protected TopNQueryResponse topN(MetadataRegistry.Schema schema,
+                                     TimestampRange timestampRange,
+                                     int number,
+                                     List<KeyValue> additionalConditions,
+                                     String[] attributes) throws IOException {
+        return topNQuery(schema, timestampRange, number, AbstractQuery.Sort.DESC, additionalConditions, attributes);
     }
 
-    protected TopNQueryResponse bottomN(MetadataRegistry.Schema schema, TimestampRange timestampRange, int number,
-                                        List<KeyValue> additionalConditions) throws IOException {
-        return topNQuery(schema, timestampRange, number, AbstractQuery.Sort.ASC, additionalConditions);
+    protected TopNQueryResponse bottomN(MetadataRegistry.Schema schema,
+                                        TimestampRange timestampRange,
+                                        int number,
+                                        List<KeyValue> additionalConditions,
+                                        String[] attributes) throws IOException {
+        return topNQuery(schema, timestampRange, number, AbstractQuery.Sort.ASC, additionalConditions, attributes);
     }
 
-    protected TopNQueryResponse topNQueryDebuggable(MetadataRegistry.Schema schema, TimestampRange timestampRange, int number,
-                                          AbstractQuery.Sort sort, List<KeyValue> additionalConditions) throws IOException {
+    protected TopNQueryResponse topNQueryDebuggable(MetadataRegistry.Schema schema,
+                                                    TimestampRange timestampRange,
+                                                    int number,
+                                                    AbstractQuery.Sort sort,
+                                                    List<KeyValue> additionalConditions,
+                                                    String[] attributes) throws IOException {
         DebuggingTraceContext traceContext = DebuggingTraceContext.TRACE_CONTEXT.get();
         DebuggingSpan span = null;
         try {
@@ -148,10 +161,12 @@ public abstract class AbstractBanyanDBDAO extends AbstractDAO<BanyanDBStorageCli
                        .append(", Sort: ")
                        .append(sort)
                        .append(", AdditionalConditions: ")
-                       .append(additionalConditions);
+                       .append(additionalConditions)
+                       .append(", Attributes: ")
+                       .append(Arrays.toString(attributes));
                 span.setMsg(builder.toString());
             }
-            TopNQueryResponse response = topNQuery(schema, timestampRange, number, sort, additionalConditions);
+            TopNQueryResponse response = topNQuery(schema, timestampRange, number, sort, additionalConditions, attributes);
             if (traceContext != null && traceContext.isDumpStorageRsp()) {
                 builder.append("\n").append(" Response: ").append(new Gson().toJson(response.getTopNLists()));
                 span.setMsg(builder.toString());
@@ -164,19 +179,32 @@ public abstract class AbstractBanyanDBDAO extends AbstractDAO<BanyanDBStorageCli
         }
     }
 
-    private TopNQueryResponse topNQuery(MetadataRegistry.Schema schema, TimestampRange timestampRange, int number,
-                                        AbstractQuery.Sort sort, List<KeyValue> additionalConditions) throws IOException {
+    private TopNQueryResponse topNQuery(MetadataRegistry.Schema schema,
+                                        TimestampRange timestampRange,
+                                        int number,
+                                        AbstractQuery.Sort sort,
+                                        List<KeyValue> additionalConditions,
+                                        String[] attributes) throws IOException {
         final TopNQuery q = new TopNQuery(List.of(schema.getMetadata().getGroup()), schema.getTopNSpec().getName(),
                 timestampRange,
                 number, sort);
         q.setAggregationType(MeasureQuery.Aggregation.Type.MEAN);
+        List<PairQueryCondition<?>> conditions = new ArrayList<>();
         if (CollectionUtils.isNotEmpty(additionalConditions)) {
-            List<PairQueryCondition<?>> conditions = new ArrayList<>(additionalConditions.size());
             for (final KeyValue kv : additionalConditions) {
                 conditions.add(PairQueryCondition.StringQueryCondition.eq(kv.getKey(), kv.getValue()));
             }
-            q.setConditions(conditions);
         }
+        if (CollectionUtils.isNotEmpty(attributes)) {
+            for (int i = 0; i < attributes.length; i++) {
+                if (StringUtil.isNotEmpty(attributes[i])) {
+                    conditions.add(
+                        PairQueryCondition.StringQueryCondition.eq(Metrics.ATTR_NAME_PREFIX + i, attributes[i]));
+                }
+            }
+        }
+        q.setConditions(conditions);
+
         return getClient().query(q);
     }